+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/utils/utils.kt
@@ -181,6 +181,7 @@ internal val Project.dependenciesHelper
         project.path,
         gradle,
         rootProjectPath,
+        extensionProvider,
         project.settings.dependencyResolutionManagement.rulesMode,
     )