+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add license information into a published POM file
 - Introduce `intellijPlatform.caching` extension
+- Allow for customizing the cached IntelliJ Platform directory name with `intellijPlatform.caching.ides.name`
 
 ### Changed